requirements.txt
# Claude OS Python Dependencies
# Install with: pip install -r requirements.txt

# === Core Framework ===
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# === Async & Database ===
aiosqlite>=0.19.0
httpx>=0.25.0
aiohttp>=3.9.0

# === Web & API ===
jinja2>=3.1.0
python-multipart>=0.0.6
sse-starlette>=1.8.0

# === Configuration ===
python-dotenv>=1.0.0
PyYAML>=6.0.0

# === File Watching ===
watchdog>=3.0.0
watchfiles>=0.21.0

# === Utilities ===
python-dateutil>=2.8.0
beautifulsoup4>=4.12.0
pytz>=2024.1
requests>=2.31.0

# === Image Generation ===
Pillow>=10.0.0

# === MCP (Model Context Protocol) ===
mcp>=0.1.0
fastmcp>=0.1.0

# === macOS Integration ===
# These provide access to Apple Calendar, Contacts, Mail
pyobjc-framework-Contacts>=10.0
pyobjc-framework-EventKit>=10.0

# === Telegram Integration ===
python-telegram-bot>=22.5

# === Optional: Cloud Services ===
# Uncomment if using Supabase for cloud sync
# supabase>=2.0.0

# === Optional: SMS Notifications ===
# Uncomment if using Twilio for SMS
# twilio>=8.0.0

# === Claude Agent SDK ===
# For programmatic Claude access (missions, scheduled tasks)
claude-agent-sdk>=0.1.0

# === Development/Testing ===
pytest>=7.0.0
pytest-timeout>=2.2.0
pytest-asyncio>=0.21.0
